refactor: move self-contained utils into utils/utils-lite
Separates crypto-free utilities from generate-random-id.ts (which imports from "crypto").
This allows consumers to import from utils/utils-lite without pulling in the crypto dependency.

We faced this issue recently when we tried to use assertNever from this lib in syntaxt-transformer
